requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.5
django-allauth==65.11.2
django-pwa==2.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
Faker==37.5.3
idna==3.10
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
